+++ src/main/java/lotto/Money.java
@@ -1,0 +1,4 @@
+package lotto;
+
+public interface Money {
+}
