+++ src/test/java/lotto/dto/WonTest.java
@@ -1,18 +1,17 @@
 package lotto.dto;
 
+import lotto.model.Won;
 import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
 import org.junit.jupiter.params.provider.NullAndEmptySource;
 import org.junit.jupiter.params.provider.ValueSource;
-import org.springframework.beans.factory.annotation.Value;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.*;
 
 @DisplayName("돈은")
-class MoneyTest {
+class WonTest {
 
 
     @ParameterizedTest
@@ -20,7 +19,7 @@ class MoneyTest {
     @DisplayName("널 또는 공백이 될 수 없습니다.")
     void moneyConstructorNullOrEmptyCheckTest(final String money){
 
-        Exception e = assertThrows(IllegalArgumentException.class, () -> new Money(money));
+        Exception e = assertThrows(IllegalArgumentException.class, () -> new Won(money));
 
         assertTrue(e.getMessage().contains("로또 구입 금액은 널 또는 공백이 될 수 없습니다."));
     }
@@ -30,7 +29,7 @@ class MoneyTest {
     @DisplayName("음수가 들어올 수 없습니다.")
     void moneyConstructorNegativeCheckTest(final String money) {
 
-        Exception e = assertThrows(IllegalArgumentException.class, () -> new Money(money));
+        Exception e = assertThrows(IllegalArgumentException.class, () -> new Won(money));
 
         assertTrue(e.getMessage().contains("로또 구입 금액은 음수가 될 수 없습니다."));
     }
@@ -44,10 +43,10 @@ class MoneyTest {
     })
     @ParameterizedTest(name = "{0}원으로 {1}원 생성")
     @DisplayName("생성이 완료되었습니다.")
-    void getMoneyTest(final String line, final Integer expectedMoney) {
+    void getWonTest(final String line, final Integer expectedWon) {
 
-        final Money money = new Money(line);
+        final Won won = new Won(line);
 
-        assertThat(money.get()).isEqualTo(expectedMoney);
+        assertThat(won.get()).isEqualTo(expectedWon);
     }
 }
